[msvc] fix ambiguous abs usage

+++ branches/avidemux_2.7/avidemux_plugins/ADM_videoFilters6/telecide/Telecide_getFrame.cpp
@@ -245,7 +245,7 @@ teleCide *_param=&configuration;
                         {
                                 // The chosen frame doesn't match the prediction.
                                 if (predicted_metric == 0) mismatch = 0.0;
-                                else mismatch = (100.0*abs(predicted_metric - lowest))/predicted_metric;
+                                else mismatch = (100.0*abs((int64_t)(predicted_metric - lowest)))/predicted_metric;
                                 if (mismatch < gthresh)
                                 {
                                         // It's close enough, so use the predicted one.
@@ -288,7 +288,7 @@ teleCide *_param=&configuration;
                                         {
                                                 // The chosen frame doesn't match the prediction.
                                                 if (predicted_metric == 0) mismatch = 0.0;
-                                                else mismatch = (100.0*abs(predicted_metric - lowest))/predicted_metric;
+                                                else mismatch = (100.0*abs((int64_t)(predicted_metric - lowest)))/predicted_metric;
                                                 if ((int) mismatch <= gthresh)
                                                 {
                                                         // It's close enough, so use the predicted one.
@@ -399,7 +399,7 @@ teleCide *_param=&configuration;
         {
                 unsigned char *dstpp, *dstpn;
                 int v1, v2, z;
-                #warning blend in place is wrong!
+//                #warning blend in place is wrong!
                 final=dst;
                 // MeanX:We should copy here as we blend from source and destination
                 // for the moment we do it in place, it is wrong.
